// WADNR.Web/src/app/shared/generated/model/program-upsert-request.ts
export class ProgramUpsertRequest { 
    ProgramName?: string | null;
    ProgramShortName?: string | null;
    OrganizationID?: number;
    ProgramPrimaryContactPersonID?: number | null;
    ProgramIsActive?: boolean;
    IsDefaultProgramForImportOnly?: boolean;
    ProgramNotes?: string | null;
    ProgramFileResourceID?: number | null;
    ProgramExampleGeospatialUploadFileResourceID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProgramUpsertRequestForm { 
    ProgramName?: FormControl<string>;
    ProgramShortName?: FormControl<string>;
    OrganizationID?: FormControl<number>;
    ProgramPrimaryContactPersonID?: FormControl<number>;
    ProgramIsActive?: FormControl<boolean>;
    IsDefaultProgramForImportOnly?: FormControl<boolean>;
    ProgramNotes?: FormControl<string>;
    ProgramFileResourceID?: FormControl<number>;
    ProgramExampleGeospatialUploadFileResourceID?: FormControl<number>;
}

export class ProgramUpsertRequestFormControls { 
    public static ProgramName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramShortName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramPrimaryContactPersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramIsActive = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsDefaultProgramForImportOnly = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramFileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProgramExampleGeospatialUploadFileResourceID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}